+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 - Fix: Update the confirmation status of send-only (no change) transactions
 
+- Unite the `History` event into `Txo{Funded,Spent}`
+
 ## 0.1.0 - 2020-05-20
 
 First release!